Correct data type of global RNG

+++ e2e_tests/tests/fixtures/keras_no_op/model_def.py
@@ -40,7 +40,7 @@ class TensorFlowRandomMetric(tf.keras.metrics.Metric):
             else:
                 return 0.0
 
-        return tf.compat.v1.py_func(my_func, [tf.ones([1], dtype=tf.float64)], tf.float64)
+        return tf.compat.v1.py_func(my_func, [tf.ones([1], dtype=tf.float32)], tf.float32)
 
 
 class NoopKerasTrial(TFKerasTrial):
